web/requirements.txt
Django==1.10.3
PyYAML==3.12
coreapi==2.0.9
django-extensions==1.7.4
django-jenkins==0.110.0
djangorestframework==3.5.2
django-rest-swagger==2.1.0
itypes==1.1.0
openapi-codec==1.1.7
requests==2.11.1
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
